+++ core/src/core/entity/statemachine/states/DefaultState.java
@@ -21,5 +21,5 @@ public class DefaultState extends EntityStateMachineNodeInterface {
   }
 
   @Override
-  public void callAction(Entity entity, float timeInState) {}
+  public void callAction(Entity entity, long timeInState) {}
 }
