+++ core/src/core/entity/statemachine/states/JumpState.java
@@ -21,5 +21,5 @@ public class JumpState extends EntityStateMachineNodeInterface {
   }
 
   @Override
-  public void callAction(Entity entity, float timeInState) {}
+  public void callAction(Entity entity, long timeInState) {}
 }
