+++ core/src/core/entity/statemachine/states/RightWalkingState.java
@@ -28,5 +28,5 @@ public class RightWalkingState extends EntityStateMachineNodeInterface {
   }
 
   @Override
-  public void callAction(Entity entity, float timeInState) {}
+  public void callAction(Entity entity, long timeInState) {}
 }
